Refactor: Mapper Class 삭제 및 dto 패키지 경로 변경
